requirements.txt
python-dotenv
streamlit
google-generativeai
Python-IO
pybase64
pillow
pymupdf